% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPLMBsplines_fit.R
\name{FPLMBsplines_fit}
\alias{FPLMBsplines_fit}
\title{FPLM fit with splines}
\usage{
FPLMBsplines_fit(y, x, u, t, freq, spl, norder, fLoss)
}
\arguments{
\item{y}{the vector of scalar responses.}

\item{x}{a matrix of the functional covariates, where each row contains the
functions evaluated on a (common) grid.}

\item{u}{the values of the explanatory variable that enters the model
non-parametrically.}

\item{t}{the grid over which the functional covariates were evaluated.}

\item{freq}{basis size for the functional regression coefficient.}

\item{spl}{basis size for the non-parametric component.}

\item{norder}{the order of the B-Splines.}

\item{fLoss}{string specifying the loss function. 'ls' for least squares,
'huang' for Huber, 'lmbrob' for MM-estimator.}
}
\value{
A list with components:
\itemize{
  \item{spl}{estimated coefficients for the non parametric term},
  \item{slope}{estimated coefficients for the functional slope},
  \item{value}{the minimum value of the loss},
  \item{scale}{the scale estimate},
  \item{slope_fun}{functional slope estimate}.
}
}
\description{
FPLM fit with splines
}
\references{
Pending.
}
