% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize.R
\name{minimize}
\alias{minimize}
\title{Minimization rutine}
\usage{
minimize(
  y,
  x_coef,
  u,
  spl_kn,
  freq,
  fLoss,
  norder,
  rob.control = lmrob.control(trace.level = 0, nResample = 5000, tuning.psi = 3.443689,
    subsampling = "simple", rel.tol = 1e-05, refine.tol = 1e-05, k.max = 2000,
    maxit.scale = 2000, max.it = 2000)
)
}
\arguments{
\item{y}{the vector of scalar responses.}

\item{x_coef}{covariate coefficients.}

\item{u}{the values of the explanatory variable that enters the model
non-parametrically.}

\item{spl_kn}{number of splines for the non-parametric term.}

\item{freq}{number of splines for the functional term.}

\item{fLoss}{string specifying the loss function. 'ls' for least squares,
'huang' for Huber, 'lmbrob' for MM-estimator.}

\item{norder}{spline order (n=4 for cubic splines)}

\item{rob.control}{optional parameters if \code{fLoss == "lmbrob"}.}
}
\value{
A vector with components:
\itemize{
  \item{spl}{estimate for the non parametric term},
  \item{slope}{estimate for the functional slope},
  \item{value}{the minimum value of the loss},
  \item{scale}{ the scale estimate.} 
}
}
\description{
This function is used internally by FPLMBsplines_fit(), not meant
for 'public' use.
}
